requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
cryptography==41.0.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
